requirements.txt
discord.py
google-generativeai==0.8.3